requirements.txt
fastapi
uvicorn
fastapi-admin
tortoise-orm
asyncpg
python-dotenv
